// src/lib/types.ts
export interface Branch {
    id: string;
    name: string;
    address: string;
    phone: string;
    isActive: boolean;
    organization_id?: string;
}